Reverted R version back to 2.12.0 for now.
PICRUSt and QIIME builds are often hanging indefinitely, and after discussing with @gregcaporaso, we decided to revert for now since we have a QIIME release on Monday. We can then look into the failures and upgrade the version after the release.

I'll do a bit more debugging on my end and post details in a separate issue.